Replace back timer outside form
