+++ src/login/components/LoginApp/Login/UseOtherDevice1.tsx
@@ -1,9 +1,11 @@
 import { fetchUseOtherDevice1, UseOtherDevice1ResponseWithQR } from "apis/eduidLogin";
+import { TimeRemainingWrapper } from "components/TimeRemaining";
 import { useAppDispatch, useAppSelector } from "login/app_init/hooks";
 import ButtonPrimary from "login/components/Buttons/ButtonPrimary";
 import loginSlice from "login/redux/slices/loginSlice";
 import React, { useEffect, useState } from "react";
 import { FormattedMessage } from "react-intl";
+import { ExpiresMeter } from "./ExpiresMeter";
 import { ResponseCodeForm, ResponseCodeValues } from "./ResponseCodeForm";
 
 /*
@@ -84,6 +86,10 @@ function RenderOtherDevice1(props: { data: UseOtherDevice1ResponseWithQR }): JSX
   const [isExpired, setIsExpired] = useState(false);
   const dispatch = useAppDispatch();
 
+  function handleTimerReachZero() {
+    setIsExpired(true);
+  }
+
   // have to pass a function to ResponseCodeForm in order for it to show the button
   function handleLoginButtonOnClick() {
     return undefined;
@@ -131,6 +137,16 @@ function RenderOtherDevice1(props: { data: UseOtherDevice1ResponseWithQR }): JSX
 
         <li>
           <FormattedMessage defaultMessage="Enter the response code shown on the other device in the form below" />
+          <div className="expiration-info device1">
+            <TimeRemainingWrapper
+              name="other-device-expires"
+              unique_id={data.display_id}
+              value={data.expires_in}
+              onReachZero={handleTimerReachZero}
+            >
+              <ExpiresMeter expires_max={data.expires_max} />
+            </TimeRemainingWrapper>
+          </div>
 
           {isExpired ? (
             <RenderFatalError error={expiredMessage} />
@@ -142,8 +158,6 @@ function RenderOtherDevice1(props: { data: UseOtherDevice1ResponseWithQR }): JSX
               handleLogin={handleLoginButtonOnClick}
               handleAbort={handleAbortButtonOnClick}
               handleSubmitCode={handleSubmitCode}
-              data={data}
-              setIsExpired={setIsExpired}
             />
           )}
         </li>
